ingestion/requirements.txt
python-dotenv
pymilvus
milvus-lite
langchain-milvus
boto3
tiktoken
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
pypdf
